move helper functions into common header
